Reduce FrameCache's memory usage
Instead of using a Range<Duration>, this uses 32 bits to store the start time and 24 bits to store the length, both in microseconds.
This reduces the size of a CachedFrame from 32 to 15 bytes, or to about 47% of its old size.
This also comes at a minor loss in precision, but that's negligible.